app/schema/serie.py
#modelo de validacao com Pydantic
from pydantic import BaseModel
from typing import Optional

class SerieSchema(BaseModel):
    titulo:str
    descricao: Optional[str] = None
    ano_lancamento: int
    class Config:
        from_attributes = True